umq: change flow control sequence number to uint8_t to save bits description: change flow control sequence number to uint8_t to save bits

+++ src/urpc/umq/umq_ub/core/private/umq_pro_ub.c
@@ -812,9 +812,9 @@ static uint32_t umq_ub_fill_fc_buf(ub_queue_t *queue, umq_buf_t **buf, umq_buf_s
     return 1;
 }
 
-static ALWAYS_INLINE uint16_t umq_ub_fc_seq_inc(uint16_t seq)
+static ALWAYS_INLINE uint8_t umq_ub_fc_seq_inc(uint8_t seq)
 {
-    uint16_t next = seq + 1;
+    uint8_t next = seq + 1;
     return (next == 0) ? 1 : next;
 }
 
@@ -869,7 +869,7 @@ static void umq_ub_fill_rx_buff_post_process(ub_queue_t *queue, umq_ub_imm_t imm
 
 static bool is_umq_ub_flow_control_msg_duplicate(ub_queue_t *real_queue, umq_ub_imm_t imm)
 {
-    uint16_t recv_seq = imm.flow_control.seq;
+    uint8_t recv_seq = imm.flow_control.seq;
     ub_flow_control_t *fc = &real_queue->flow_control;
 
     switch (imm.bs.type) {
